Replace connect from index.js to Home.jsx Add constants for action (placed at src/store/constants/actions)
