+++ src/Pages/Home/style.css
@@ -1,0 +1,18 @@
+.home-page__wrapper{
+    padding: 10px;
+}
+
+.home-page__button{
+    padding: 5px 20px;
+    margin: 0 5px;
+    border-radius: 10px;
+    border: none;
+    outline: none;
+    background-color: #b9b1de;
+}
+
+.home-page__content{
+    border-radius: 20px;
+    outline: none;
+    text-indent: 5px;
+}
