+++ src/store/actions/counter.js
@@ -1,5 +1,7 @@
-export const INCREASE_COUNTER = 'INCREASE_COUNTER';
-export const DECREASE_COUNTER = 'DECREASE_COUNTER';
+import {
+  INCREASE_COUNTER,
+  DECREASE_COUNTER
+} from '../constants/actions'
 
 export function increaseCounter(increment) {
   return {
