+++ src/store/constants/actions.js
@@ -1,0 +1,2 @@
+export const INCREASE_COUNTER = 'INCREASE_COUNTER';
+export const DECREASE_COUNTER = 'DECREASE_COUNTER';
